Return ragged data (#36)
This is a major breaking change for GVL. Users should view the "What's a gvl.Dataset?" page in the documentation for details, but major breaks include:

- removed the `length` argument from gvl.write(). Regions/BED files are now used as-is. If you want uniform length regions centered on inputs/peaks as before, preprocess your BED file with `gvl.with_length`.
- changed Dataset.output_length from a property to a dynamic setting with behavior describe in the "What's a gvl.Dataset?" page.
- changed track output shape to have a track axis.
- Datasets are now deterministic by default.

As a result of these changes, GVL seamlessly supports ragged length output and also paves the way for on-the-fly splicing. Since many changes were made, I wouldn't be surprised if a few bugs crop up despite my best efforts -- please leave issues if so!

* feat: option to return ragged data from gvl.Dataset

* fix: disable shifts for ragged output. fix(wip): may need to skip some variants if they are outside ragged region.

* feat(wip): ragged output

* fix: incorrect mask from get_keep_mask_for_length (#37)

* bump: version 0.8.0 → 0.8.1

* feat!: output_length is set dynamically. fix: hap reconstruction matches bcftools

* feat!: change default for Dataset.deterministic from False to True. change track output from a list of arrays to having a track dimension i.e. from shape (b [p] l) to (b t [p] l). docs: add dataset.md, faq.md and overhaul geuvadis.ipynb to be simpler and reflect changes in API.

---------

+++ python/genvarloader/_variants/_records.py
@@ -80,7 +80,7 @@ class VLenAlleles:
 
         # handle empty result
         if start >= len(self) or (stop is not None and stop <= start):
-            return VLenAlleles(np.empty(0, np.uint64), np.empty(0, "|S1"))
+            return VLenAlleles(np.empty(0, np.int64), np.empty(0, "|S1"))
 
         if stop is None:
             stop = len(self)
@@ -183,14 +183,14 @@ class RecordInfo:
             positions=np.empty(n_queries, np.int32),
             size_diffs=np.empty(n_queries, np.int32),
             refs=VLenAlleles(
-                np.zeros(n_queries + 1, np.uint64), np.empty(n_queries, "S1")
+                np.zeros(n_queries + 1, np.int64), np.empty(n_queries, "S1")
             ),
             alts=VLenAlleles(
-                np.zeros(n_queries + 1, np.uint64), np.empty(n_queries, "S1")
+                np.zeros(n_queries + 1, np.int64), np.empty(n_queries, "S1")
             ),
             start_idxs=np.empty(n_queries, np.int32),
             end_idxs=np.empty(n_queries, np.int32),
-            offsets=np.zeros(n_queries + 1, np.int32),
+            offsets=np.zeros(n_queries + 1, np.int64),
         )
 
     @staticmethod
